+++ ui/js/component/publishForm/index.js
@@ -2,4 +2,4 @@ import React from "react";
 import { connect } from "react-redux";
 import PublishForm from "./view";
 
-export default connect()(PublishForm);
+export default connect(null, null)(PublishForm);